--- include/debug.h
@@ -1,14 +1,0 @@
-#ifndef DEBUG_H
-#define DEBUG_H
-
-
-class debug
-{
-    public:
-        debug();
-        virtual ~debug();
-    protected:
-    private:
-};
-
-#endif // DEBUG_H
