--- main.cpp
@@ -1,9 +1,0 @@
-#include <iostream>
-
-using namespace std;
-
-int main()
-{
-    cout << "Hello world!" << endl;
-    return 0;
-}
