+++ src/c_logger.h
