+++ src/galeshapley.h
@@ -1,0 +1,7 @@
+#ifndef galeshapley_h
+#define galeshapley_h
+
+List galeShapleyMatching(umat& prefM, mat& rankW);
+bool checkStability(mat& uM, mat& uW, umat& proposals, umat& engagements);
+
+#endif
