--- src/main.h
@@ -1,9 +1,0 @@
-#ifndef main_h
-#define main_h
-
-List galeShapleyMatching(umat prefM, mat rankW);
-umat sortIndex(mat u);
-umat rankIndex(umat sortedIdx);
-bool checkStability(mat uM, mat uW, umat proposals, umat engagements);
-
-#endif
