+++ src/roommate.h
@@ -1,5 +1,5 @@
-#ifndef stable_h
-#define stable_h
+#ifndef roommate_h
+#define roommate_h
 
 List stableRoommateMatching(const umat pref);
 void deleteValueWithWarning(std::vector<size_t> *vec, size_t val);
