+++ src/utils.h
@@ -1,0 +1,7 @@
+#ifndef utils_h
+#define utils_h
+
+umat sortIndex(const mat& u);
+umat rankIndex(const umat& sortedIdx);
+
+#endif
